requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
DateTime==4.3
Django==3.1.4
django-redis==4.12.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mysql-client==0.0.1
mysqlclient==2.0.2
PyJWT==1.7.1
pytz==2020.5
redis==3.5.3
requests==2.25.1
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.2
XlsxWriter==1.3.7
zope.interface==5.2.0
